+++ application/shared/view/timeclock/admin/about.php
@@ -1,10 +1,20 @@
 <div class="jumbotron">
     <h1>TimeClock</h1>
     <div class="row">
-        <div class="col-sm-12">
+        <div class="col-sm-8">
             <div class="well">
                 <p>This system is designed as a management panel for external timeclock hardware.</p>
                 <a href="{timeclock_root}about" class="btn btn-primary btn-lg" role="button">Learn more &raquo;</a>
+            </div>
+        </div>
+        <div class="col-sm-4">
+            <div class="panel panel-primary">
+                <div class="panel-heading">
+                    <h3 class="panel-title">Employee Mangement</h3>
+                </div>
+                <div class="panel-body center">
+                    <a href="{timeclock_root}employee/add">Add New Employee</a>
+                </div>
             </div>
         </div>
     </div>