+++ application/shared/view/timeclock/admin/employee/view.php
@@ -54,6 +54,7 @@ $date_format = 'm/d/y';
                             </tr>
                         </tfoot>
                     </table>
+                    <a href="javascript:addDate({pay_period_id})">Add Date</a>
                 </div>
             </div>
             <div class="row">
@@ -141,3 +142,33 @@ $date_format = 'm/d/y';
         </div> <!-- END: well -->
     </div> <!-- END: col-sm-12 -->
 </div> <!-- END: row -->
+<div class="update_time_dialog">
+    <form class="update_time_form" method="post" action="">
+        <div class="dialog_text">
+            <div class="bold dialog_title"></div>
+            <input class="dialog_input" type="text" name="time" value="3:00pm" />
+        </div>
+        <input type="hidden" name="employee_id" value="{employees_by_id[<?php echo $this->sys->template->employee_id ?>]['id']}" />
+        <input type="hidden" name="date" />
+        <input type="hidden" name="time_index" />
+        <input type="hidden" name="time_operation" />
+        <input type="hidden" name="update_time" value="update" />
+    </form>
+</div>
+<div class="add_date_dialog">
+    <form class="add_date_form" method="post" action="">
+        <div class="dialog_text">
+            <div class="date"></div>
+            <input class="date_to_add" type="hidden" name="date" />
+        </div>
+        <input type="hidden" name="employee_id" value="{employees_by_id[<?php echo $this->sys->template->employee_id ?>]['id']}" />
+        <input type="hidden" name="id" />
+        <input type="hidden" name="add_date" value="add_date" />
+    </form>
+</div>
+<div class="add_date_response">
+    <div class="add_date_response_text">
+        All previous in/out slots must be filled
+    </div>
+</div>
+{add_date_response}