+++ application/shared/view/timeclock/admin/home.php
@@ -1,10 +1,20 @@
 <div class="jumbotron">
     <h1>TimeClock</h1>
     <div class="row">
-        <div class="col-sm-12">
+        <div class="col-sm-8">
             <div class="well">
                 <p>This system is designed as a management panel for external timeclock hardware.</p>
                 <a href="{timeclock_root}about" class="btn btn-primary btn-lg" role="button">Learn more &raquo;</a>
+            </div>
+        </div>
+        <div class="col-sm-4">
+            <div class="panel panel-primary">
+                <div class="panel-heading">
+                    <h3 class="panel-title">Employee Mangement</h3>
+                </div>
+                <div class="panel-body center">
+                    <a href="{timeclock_root}employee/add">Add New Employee</a>
+                </div>
             </div>
         </div>
     </div>
@@ -25,6 +35,7 @@
                             <th>Role</th>
                             <th>Category</th>
                             <th>Username</th>
+                            <th>Manage</th>
                         </tr>
                     </thead>
                     <tbody>
@@ -42,6 +53,12 @@
                             <td onclick="employeeTableClicked('view', '{employees[<?php echo $i; ?>]['id']}')">{employees[<?php echo $i; ?>]['role']}</td>
                             <td onclick="employeeTableClicked('view', '{employees[<?php echo $i; ?>]['id']}')">{employees[<?php echo $i; ?>]['category_name']}</td>
                             <td onclick="employeeTableClicked('view', '{employees[<?php echo $i; ?>]['id']}')">{employees[<?php echo $i; ?>]['username']}</td>
+                                <td>
+                                    <ul class="icons">
+                                    <li class="ui-state-default ui-corner-all" title=".ui-icon-pencil"><span class="ui-icon ui-icon-pencil" onclick="employeeTableClicked('edit', '{employees[<?php echo $i; ?>]['id']}')"></span></li>
+                                    <li class="ui-state-default ui-corner-all" title=".ui-icon-trash"><span class="ui-icon ui-icon-trash" onclick="employeeTableClicked('trash', '{employees[<?php echo $i; ?>]['id']}')"></span></li>
+                                    </ul>
+                                </td>
                             </tr>
                             <?php
                         }
@@ -55,6 +72,7 @@
                             <th>Role</th>
                             <th>Category</th>
                             <th>Username</th>
+                            <th>Manage</th>
                         </tr>
                     </tfoot>
                 </table>