+++ application/shared/view/timeclock/includes/navbar-admin.php
@@ -14,6 +14,7 @@
         <li {home_active}><a href="{timeclock_root}main">Home</a></li>
         <li {jobs_active}><a href="{timeclock_root}jobs">Jobs</a></li>
         <li {about_active}><a href="{timeclock_root}about">About</a></li>
+        <li {settings_active}><a href="{timeclock_root}settings">Settings</a></li>
         <li><a href="{timeclock_root}logout">Logout</a></l>
       </ul>
     </div><!--/.nav-collapse -->